+++ actionview/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 7.1.1 (October 11, 2023) ##
+
 *   Updated `@rails/ujs` files to ignore certain data-* attributes when element is contenteditable.
 
     This fix was already landed in >= 7.0.4.3, < 7.1.0.